refactor: Split into multiple files
- Added registry, setups, and teardowns to control plugin actions
