+++ src/pbr_preview/src/deps.ts
@@ -1,0 +1,5 @@
+// @ts-expect-error THREE UMD is OK
+const three = THREE;
+const bb = Blockbench;
+
+export { three, bb };
