+++ src/pbr_preview/src/lib/actions/index.ts
@@ -1,0 +1,9 @@
+import "./bakeTextures";
+import "./channelAssignment";
+import "./generateNormal";
+import "./materialTexture";
+import "./mer";
+import "./textureSet";
+import "./toggleLights";
+import "./togglePbr";
+import "./tonemapping";
